Added RLE
Changed file name and added RLE encoding and decoding.
